Outsource error formatting to SemanticError
